==> R/00_all.qmd <==
---
title: "00_all"
format: html
editor: visual
---

# Packages

```{r}
#| message: false
library("tidyverse")
library("dplyr")
library("stringr")
library("tidyr")
library("readxl")
library("purrr")
```

# Creating directories

```{r}
createOrCheckDirectory <- function(directory) {
  if (!file.exists(directory)) {
    dir.create(directory, recursive = TRUE)
    cat(paste("Directory '", directory, "' created.\n", sep = ""))
  } else {
    cat(paste("Directory '", directory, "' already exists.\n", sep = ""))
  }
}
createOrCheckDirectory("../data/_raw")
createOrCheckDirectory("../results")
createOrCheckDirectory("../doc")
```

```{r}
rmarkdown::render("01_load.qmd", output_dir = "../results/")

rmarkdown::render("02_clean.qmd", output_dir = "../results/")

rmarkdown::render("03_augment.qmd", output_dir = "../results/")

rmarkdown::render("04_describe.qmd", output_dir = "../results/")

rmarkdown::render("05_analysis_1.qmd", output_dir = "../results/")

rmarkdown::render("06_analysis_2.qmd", output_dir = "../results/")
```
