==> R/01_load.qmd <==
```{r}
#| label: Loading packages
library(tidyverse)
library(readr)
library(readxl)
```

```{r}
#| label: Downloading Byar & Greene prostate cancer data
url <- "https://github.com/Alex131313131313/DATA/raw/main/prostate.xlsx"
destination <- "../data/_raw/prostate.xlsx"

download.file(url, 
              destination, 
              method = "auto")
```

```{r}
#| label: Converting data from xlsx to tsv
# NOTE: You open prostate.xls and save it again to xlsx
prostate_data <- read_excel('../data/_raw/prostate.xlsx')
prostate_data <- write_delim(prostate_data, 
                             file= '../data/_raw/prostate.tsv',      
                             delim = "\t", col_names = TRUE)
```

```{r}
#|label: Saving data

output_file_path <- "../data/01_dat_load.tsv"
write_delim( prostate_data,
             file = output_file_path )
```
