==> R/02_clean.qmd <==
```{r}
#| label: Loading libraries
#| message: false

library("tidyverse")
library("patchwork") #for plotting plots together (plo1/plot2)
#library("viridis")
library("dplyr")
library("stringr") #count the number of occurrences of a given character in a string
library("readr")
```

```{r}
#| label: Loading data
prostate_data <- read_delim(file = "../data/01_dat_load.tsv", 
                            show_col_types = FALSE)
```

```{r}
#| label: Renaming rx column, converting to numeric, replacing NAs 
#| warning: false

prostate_data_clean <- prostate_data |>
  separate("rx", 
           into = c("Treatment regime",
                    "mg",
                    "Drug"), 
           sep =" ") |>
  select(-c("mg"))


prostate_data_clean <- prostate_data_clean |>
  mutate(`Treatment regime` = case_when(
    `Treatment regime` == "placebo" ~ "0",
    TRUE ~ as.character(`Treatment regime`))) |>
  mutate(`Treatment regime` = as.numeric(`Treatment regime`))
 
prostate_data_clean <- prostate_data_clean |>
  mutate(`Drug` = case_when(
    `Drug` == "estrogen" ~ "1",
    TRUE ~ as.character(`Drug`))) |>
  mutate(`Drug` = as.numeric(`Drug`)) |>
  mutate(`Drug` = replace_na(`Drug`,0))

```



```{r}
#| label: Modifying Status column, creating new columns, converting str to numeric
#| warning: false

prostate_data_clean <- prostate_data_clean |>
  separate("status", 
           into = c("Status", 
                    "Cause of Death"), 
           sep ="\\-")

prostate_data_clean <- prostate_data_clean |>
  mutate(`Status` = case_when(
    `Status` == "alive" ~ "1",
    TRUE ~ as.character(`Status`))) |>
  mutate(`Status` = as.numeric(`Status`)) |>
  mutate(`Status` = replace_na(`Status`,0))
```

```{r}
#| label: Modifying PF column

prostate_data_clean <- prostate_data_clean |>
  mutate(`pf` = if_else(
    `pf` == "normal activity", 0,1))
  
```

```{r}
#| label: Renaming columns appropriately
prostate_data_clean <- prostate_data_clean |>
  rename(`Activity Level` = pf,
         `Patient number` = `patno`,
         `Systolic blood pressure` = `sbp`,
         `Diastolic blood pressure` = `dbp`,
         `Serum haemoglobin level` = `hg`,
         `Weight index` = `wt`,
         `History of CVD` = `hx`,
         `Primary lesion size(cm^2)` = `sz`,
         `Gleason stage` = `sg`) 
```


```{r}
#| label: Converting 'Cause of Death' column to numeric values 

prostate_data_clean <- prostate_data_clean |>
  mutate(`Cause of Death` = case_when(
    `Cause of Death` == " prostatic ca" ~ "0",
    `Cause of Death` == " heart or vascular" ~ "1",
    `Cause of Death` == " cerebrovascular" ~ "2",
    `Cause of Death` == " other ca" ~ "3",
    `Cause of Death` == " respiratory disease" ~ "4",
    `Cause of Death` == " other specific non" ~ "5",
    `Cause of Death` == " pulmonary embolus" ~ "6",
    `Cause of Death` == " unspecified non" ~ "7",
    `Cause of Death` == " unknown cause" ~ "8",
    TRUE ~ as.character(`Cause of Death`))) |>
  mutate(`Cause of Death` = as.numeric(`Cause of Death`)) |>
  mutate(`Cause of Death` = replace_na(`Cause of Death`,9))
```

```{r}
#|label: Save data

output_file_path <- "../data/02_dat_clean.tsv"
write_delim( prostate_data_clean, 
             file = output_file_path )

```
