==> R/05_analysis_1.qmd <==
---
title: "05_analysis_1"
format: html
editor: visual
---

```{r}
library(purrr)
library(dplyr)
library(tidyr)
library(broom)
```

```{r}
#Filter out NAs in the variables we are interested in
prostate_data_analysis_01 <- prostate_data_augment %>% 
  mutate(stage = factor(stage)) %>% 
  mutate(Dosage = factor(Dosage)) %>% 
  filter(!is.na(`bm`),
         !is.na(`Weight index`),
         !is.na(`Status`),
         !is.na(`Primary lesion size(cm^2)`),
         !is.na(`Age Adjusted Haemoglobin`) ,
         !is.na(`Dosage`))%>% 
  group_by(stage) %>% 
  nest() %>% 
  ungroup()

# Fit logistic linear regression to model Dosage based on other variables for each stage
prostate_data_analysis_01 <- prostate_data_analysis_01 %>%
  mutate(model_object = map(.x = data,
      .f = ~ glm(`Dosage` ~ `bm` +
                   `Weight index` +
                   `Primary lesion size(cm^2)` +
                   `Age Adjusted Haemoglobin`,
            family = binomial(link = "logit"),
            data = .x )))

  
# Extract coefficients from the glm objects within each nested dataframe
prostate_estimates <- prostate_data_analysis_01 %>%
  mutate(
    model_summary = map(model_object, 
                        tidy)
  ) %>%
  unnest(model_summary)

# Create a new variable to flag p-values less than 0.05
prostate_estimates <- prostate_estimates %>%
  mutate(significant = ifelse(p.value < 0.05,
                              "Yes",
                              "No"))

```

```{r}
# Plotting coefficients
plt_glm <- ggplot(prostate_estimates, aes(x = term, 
                                          y = estimate, 
                                          group = stage,  
                                          color = stage, 
                                          shape = significant )) +
  geom_point() +
  theme_minimal() +
  labs(x = "Coefficients",
       y = "Estimate") +
  theme(axis.text.x = element_text(angle=45, 
                                   hjust=1)) +
  ggtitle("Logistic Regression Coefficients to predict Dosage by Stage")

plt_glm

#Save
ggsave('../results/glm_01.png', plt_glm)
```

```{r}
plt_glm <- ggplot(prostate_estimates, aes(x = term, y = estimate, fill = stage, shape = significant)) +
  geom_bar(stat = "identity", position = "dodge") +
  geom_text(aes(label = significant), 
            position = position_dodge(width = 0.9), 
            vjust = -0.5, 
            size = 3) +
  theme_minimal() +
  labs(x = "Coefficients", y = "Estimate") +
  theme(axis.text.x = element_text(angle = 45, 
                                   hjust = 1)) +
  ylim( -20, 5 )
  ggtitle("Logistic Regression Coefficients to Predict Dosage by Stage")

plt_glm

#Save
ggsave('../results/glm_02.png', plt_glm)
```
