==> R/06_analysis_2.qmd <==
---
title: "06_analysis_2"
format: html
editor: visual
---

```{r}
library(broom)
library(scales)
```

```{r}
variables_to_plot <- prostate_data_augment[, c("Primary lesion size(cm^2)",
                                               "age", 
                                               "Serum haemoglobin level", 
                                               "Gleason stage")]

# Reshape the data for ggplot
data_long <- gather(variables_to_plot, 
                    key = "Variable", 
                    value = "Value")

# Create multiple Q-Q plots with facet_wrap
ggplot(data_long,
       aes(sample = Value)) +
  geom_qq() +
  facet_wrap(~ Variable, scales = "free") +
  ggtitle("Multiple Q-Q Plots")
```

```{r}
prostate_data_augment <-prostate_data_augment |>
  filter(!is.na(age)) |>
  filter(!is.na(`Primary lesion size(cm^2)`)) |>
  filter(!is.na(`Gleason stage`)) |>
   filter(!is.na(`Weight index`))
```

```{r}

#method 1 - normalise data and standardise data

pca_fit <- prostate_data_augment |>
  select(`Treatment regime`,
         `age`, 
         `Primary lesion size(cm^2)`,
         `Gleason stage`, 
         `Serum haemoglobin level`, 
         `stage`,
         `History of CVD`,
         `Activity Level`,
         `bm`)|>
  prcomp(scale=TRUE) 

```

```{r}
pca_fit|>
  augment(prostate_data_augment)|> # add original dataset back in
  ggplot(aes(.fittedPC1,
             .fittedPC2, 
             color = stage)) + 
  geom_point(size = 1.5) +
  theme_minimal() +
    ggtitle("PCA analysis coloured by Stage") +
  labs(x = "Fitted Principal component 1",
       y = "Fitted Principal component 2" )


ggsave(filename = '../results/pca1.png')
```

```{r}
pca_fit |>
  augment(prostate_data_augment) %>% # add original dataset back in
  ggplot(aes(.fittedPC1, 
             .fittedPC2,
             color = `Primary lesion size(cm^2)`)) + 
  geom_point(size = 1.5) +
  theme_minimal() +
    ggtitle("PCA analysis coloured by Primary Lesion size")+
  labs(x = "Fitted Principal component 1",
       y = "Fitted Principal component 2" )

ggsave(filename = '../results/pca2.png')
```

```{r}
# extracting the rotation matrix
pca_fit |>
  tidy(matrix = "rotation")

```

```{r}
arrow_style <- arrow(
  angle = 20, ends = "first", type = "closed", length = grid::unit(8, "pt"))

# plot rotation matrix
pca_fit %>%
  tidy(matrix = "rotation") %>%
  pivot_wider(names_from = "PC", 
              names_prefix = "PC", 
              values_from = "value") %>%
  ggplot(aes(PC1, PC2)) +
  geom_segment(xend = 0,
               yend = 0,
               arrow = arrow_style) +
  geom_text(
    aes(label = column),
    hjust = 1, 
    nudge_x = -0.02, 
    color = "skyblue2") +
  xlim(-1.25, .5) +
  ylim(-.5, 1) +
  coord_fixed() +
  ggtitle("PCA Loading Plot")+
  labs(x = "Fitted Principal component 1",
       y = "Fitted Principal component 2" )

ggsave(filename = '../results/pca_arrows.png')
```

```{r}
pca_fit %>%
  tidy(matrix = "eigenvalues")

pca_fit %>%
  tidy(matrix = "eigenvalues") %>%
  ggplot(aes(PC, percent)) +
  geom_col(fill = "#56B4E9",
           alpha = 0.8) +
  scale_x_continuous(breaks = 1:9) +
  scale_y_continuous(
    labels = scales::percent_format(),
    expand = expansion(mult = c(0, 0.01))
  )  +
  labs(title = "Variance explained by each principal component",
       x = "Principal component", 
       y = "% Variance")

ggsave(filename = '../results/pca1_bars.png')
```
